Delete WelkinSuite.munki.recipe
Removing munki recipe as the preinstall and postinstall scripts for The_Welkin_Suite.pkg hardcode the current user's home folder, and as munki runs as root, this will not work to my knowledge. Emailed developer requesting they add logic to get the current logged in user instead of using ~/ in their scripts.